% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_path.R
\name{capac_info}
\alias{capac_info}
\title{capac_info}
\usage{
capac_info(net, lambda, eta, silent = 0)
}
\arguments{
\item{net}{a gene network}

\item{lambda}{lambda value}

\item{eta}{eta}

\item{silent}{silent}
}
\description{
capac_info
}
