% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_scored_net.R
\name{construct_scored_net}
\alias{construct_scored_net}
\title{construct_scored_net}
\usage{
construct_scored_net(
  gene_network_data,
  interaction_indices,
  weight_index = NULL,
  gene_ps,
  gene_scores,
  genes2exclude = NULL
)
}
\arguments{
\item{gene_network_data}{a gene network}

\item{interaction_indices}{indices of interaction}

\item{weight_index}{weight_index}

\item{gene_ps}{gene_ps}

\item{gene_scores}{gene_scores}

\item{genes2exclude}{(optional) genes to exclude for analysis}
}
\description{
construct_scored_net
}
