% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_string_net.R
\name{construct_string_net}
\alias{construct_string_net}
\title{construct_string_net}
\usage{
construct_string_net(
  gene_ps,
  gene_scores,
  evidence2exclude = c("textmining_transferred", "textmining"),
  genes2exclude = NULL,
  confidence = 0.7
)
}
\arguments{
\item{gene_ps}{gene level p-values}

\item{gene_scores}{gene scores}

\item{evidence2exclude}{which evidence to exclude}

\item{genes2exclude}{which genes to exclude}

\item{confidence}{confidence cutoff}
}
\description{
construct_string_net
}
