% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_selected_p_values.R
\name{hist_selected_p_values}
\alias{hist_selected_p_values}
\title{hist_selected_p_values}
\usage{
hist_selected_p_values(whole_ps, selected_ps, output)
}
\arguments{
\item{whole_ps}{all p-values (intermediate file returned from helper function)}

\item{selected_ps}{selected p-values}

\item{output}{output pdf file}
}
\description{
hist_selected_p_values
}
