% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_result_table.R
\name{obtain_results}
\alias{obtain_results}
\title{obtain_results}
\usage{
obtain_results(computation_details, net)
}
\arguments{
\item{computation_details}{computation_details (intermediate file returned from helper function)}

\item{net}{network}
}
\description{
obtain_results
}
