% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_gene_family_info.R
\name{OR_gene_family_info}
\alias{OR_gene_family_info}
\title{OR_gene_family_info}
\usage{
OR_gene_family_info(save_to_local = FALSE)
}
\arguments{
\item{save_to_local}{whether to save the result to local}
}
\description{
OR_gene_family_info
}
