% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2score.R
\name{p2score}
\alias{p2score}
\title{p2score
converts p-values to upper quantile of normal distribution}
\usage{
p2score(gene_ps)
}
\arguments{
\item{gene_ps}{gene p-values as data.frame}
}
\value{
gene_scores
}
\description{
p2score
converts p-values to upper quantile of normal distribution
}
