% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_fw.R
\name{path_fw}
\alias{path_fw}
\title{path_fw}
\usage{
path_fw(net, lambda, edge_limit)
}
\arguments{
\item{net}{a gene network}

\item{lambda}{lambda value}

\item{edge_limit}{edge_limit}
}
\description{
path_fw
}
