% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_fw_weighted.R
\name{path_fw_weighted}
\alias{path_fw_weighted}
\title{SigMod_bisection}
\usage{
path_fw_weighted(net, lambda, edge_limit)
}
\arguments{
\item{net}{a gene network}

\item{edge_limit}{edge_limit}

\item{lamba}{lambda value}
}
\description{
SigMod_bisection
}
