% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_path.R
\name{reg_path}
\alias{reg_path}
\title{reg_path}
\usage{
reg_path(
  net,
  lambda,
  eta_max,
  eta_min,
  capac_infos,
  selection_diff = 1,
  selection_min,
  selection_max,
  silent
)
}
\arguments{
\item{net}{a gene network}

\item{lambda}{lambda value}

\item{eta_max}{eta_max}

\item{eta_min}{eta_min}

\item{capac_infos}{capac_infos}

\item{selection_diff}{selection_diff}

\item{selection_min}{selection_min}

\item{selection_max}{selection_max}

\item{silent}{silent}
}
\description{
reg_path
}
