% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_path.R
\name{selection_path}
\alias{selection_path}
\title{selection_path}
\usage{
selection_path(
  net,
  lambda,
  eta_max,
  eta_min,
  selection_diff,
  selection_min,
  selection_max,
  silent
)
}
\arguments{
\item{net}{a gene network}

\item{lambda}{lambda value}

\item{eta_max}{eta_max}

\item{eta_min}{eta_min}

\item{selection_diff}{selection_diff}

\item{selection_min}{selection_min}

\item{selection_max}{selection_max}

\item{silent}{silent}
}
\description{
selection_path
}
