% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_path_fast.R
\name{selection_path_fast}
\alias{selection_path_fast}
\title{selection_path_fast}
\usage{
selection_path_fast(
  net,
  lambda,
  eta_max = 1e+07,
  eta_min,
  selection_diff = 1,
  selection_min = 0,
  selection_max = Inf,
  silent
)
}
\arguments{
\item{net}{a gene network}

\item{lambda}{penalty value}

\item{eta_max}{eta_max}

\item{eta_min}{eta_min}

\item{selection_diff}{selection_diff}

\item{selection_min}{selection_min}

\item{selection_max}{selection_max}

\item{silent}{silent}
}
\description{
selection_path_fast
}
