% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigMod_bisection.R
\name{SigMod_bisection}
\alias{SigMod_bisection}
\title{SigMod_bisection}
\usage{
SigMod_bisection(
  net,
  lambda_min = 0,
  lambda_max = 1,
  eta_max = 1e+07,
  size_diff = 1,
  size_min = 0,
  nmax = 300,
  maxjump = 30,
  epsilon = 1e-05,
  cp = 1000,
  silent = 1
)
}
\arguments{
\item{net}{a gene network}

\item{lambda_min}{minimum lambda value}

\item{lambda_max}{max lambda value}

\item{eta_max}{eta_max}

\item{size_diff}{size_diff}

\item{size_min}{size_min}

\item{nmax}{nmax}

\item{maxjump}{maxjump}

\item{epsilon}{epsilon}

\item{cp}{cp}

\item{silent}{silent}
}
\description{
SigMod_bisection
}
