% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigMod_core.R
\name{SigMod_core}
\alias{SigMod_core}
\title{SigMod_core}
\usage{
SigMod_core(net, lambda, eta, silent = 0)
}
\arguments{
\item{net}{a gene network}

\item{lambda}{lambda value}

\item{eta}{eta}

\item{silent}{silent}
}
\description{
SigMod_core
}
